// iheart-songs/src/API.ts
/* tslint:disable */
/* eslint-disable */
//  This file was automatically generated and should not be edited.

export type Song = {
  __typename: "Song",
  song?: string | null,
  artist?: string | null,
  songReleaseDate?: string | null,
  playCount?: number | null,
  metricA?: number | null,
  metricB?: number | null,
  metricC?: number | null,
  metricD?: number | null,
  metricE?: number | null,
  metricF?: number | null,
  metricG?: number | null,
  metricH?: number | null,
  metricI?: number | null,
  metricJ?: number | null,
  metricK?: number | null,
  metricL?: number | null,
  metricM?: number | null,
  metricN?: number | null,
  metricO?: number | null,
  metricP?: number | null,
};

export type GetSongsQuery = {
  getSongs?:  Array< {
    __typename: "Song",
    song?: string | null,
    artist?: string | null,
    songReleaseDate?: string | null,
    playCount?: number | null,
    metricA?: number | null,
    metricB?: number | null,
    metricC?: number | null,
    metricD?: number | null,
    metricE?: number | null,
    metricF?: number | null,
    metricG?: number | null,
    metricH?: number | null,
    metricI?: number | null,
    metricJ?: number | null,
    metricK?: number | null,
    metricL?: number | null,
    metricM?: number | null,
    metricN?: number | null,
    metricO?: number | null,
    metricP?: number | null,
  } | null > | null,
};